requirements.txt
fastapi
uvicorn
python-dotenv
ics
cloud-sql-python-connector
sqlalchemy
psycopg2